+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleProjectDAO.scala
@@ -12,4 +12,6 @@ trait GoogleProjectDAO {
 
   def isProjectActive(projectName: String): Future[Boolean]
 
+  def isBillingActive(projectName: String): Future[Boolean]
+
 }