+++ google/src/test/scala/org/broadinstitute/dsde/workbench/google/mock/MockGoogleProjectDAO.scala
@@ -13,4 +13,6 @@ class MockGoogleProjectDAO extends GoogleProjectDAO {
   override def pollOperation(operationId: String): Future[Operation] = Future.successful(new Operation)
 
   override def isProjectActive(projectName: String): Future[Boolean] = Future.successful(true)
+
+  override def isBillingActive(projectName: String): Future[Boolean] = Future.successful(true)
 }